fix(vu): correctly handle 128-bit device features
The vhost-user protocol requires 64-bit device features, but the virtio
spec allows 128. This commit truncates features to 64 bits before
sending, fixing a bug where all 128 bits were sent, causing feature
negotiation to fail.

Fixes: 128153e1e017 ("fix(virtio): extend virtio feature to 128 bits")

+++ alioth/src/virtio/vu/backend.rs
@@ -283,7 +283,7 @@ impl VuBackend {
             }
             (VuFrontMsg::GET_FEATURES, 0) => {
                 let feature = self.dev.device_feature | VirtioFeature::VHOST_PROTOCOL.bits();
-                self.session.reply(req, &feature, &[])?;
+                self.session.reply(req, &(feature as u64), &[])?;
                 msg.flag.set_need_reply(false);
                 log::debug!("{name}: get device feature: {feature:#x}");
             }
